requirements.txt
streamlit
sqlite3
python-dotenv
plotly
bcrypt
